+++ app/controllers/admin_controller.rb
@@ -1,14 +1,8 @@
 class AdminController < ApplicationController
   before_filter :authenticate_user!
 
-  # ensure_authorization_performed
-  #
-  # authorize_actions_for User, actions: { dashboard: :update, add_role_to_user: :update, remove_role_from_user: :update }
+  authorize_actions_for User, only: [:add_role_to_user, :remove_role_from_user, :destroy_user], actions: {add_role_to_user: :update, remove_role_from_user: :update, destroy_user: :delete}
 
-  # User and Survey are focuses now
-  # For user, we allow owner to update users
-
-  authorize_actions_for User, only: [:add_role_to_user, :remove_role_from_user], actions: {add_role_to_user: :update, remove_role_from_user: :update}
 
   def dashboard
     if current_user.can?(:view_admin_dashboard)
@@ -43,10 +37,20 @@ class AdminController < ApplicationController
     render "users"
   end
 
+  def destroy_user
+    user = User.find(params[:user_id])
+
+    user.destroy unless user == current_user
+
+    set_users
+
+    render "users"
+  end
+
   private
 
   def set_users
-    @users = (params[:search] ? User.search_by_email(params[:search]) : User.all)
+    @users = User.scoped_users(params[:search], params[:search_role])
   end