+++ app/models/user.rb
@@ -7,7 +7,6 @@ class User < ActiveRecord::Base
 
   self.authorizer_name = "UserAuthorizer"
 
-
   # Include default devise modules. Others available are:
   # :confirmable, :lockable, :timeoutable and :omniauthable
   devise :database_authenticatable, :registerable,
@@ -16,9 +15,18 @@ class User < ActiveRecord::Base
   has_many :answer_sessions
   has_many :answers
   has_many :votes
-
+  has_one :social_profile
 
   scope :search_by_email, ->(terms) { where("LOWER(#{self.table_name}.email) LIKE ?", terms.to_s.downcase.gsub(/^| |$/, '%')) }
+
+  def self.scoped_users(email=nil, role=nil)
+    users = all
+
+    users = users.search_by_email(email) if email.present?
+    users = users.with_role(role) if role.present?
+
+    users
+  end
 
   def forem_name
     email